+++ plugins/dotplot-view/src/DotplotView/components/Axes.tsx
@@ -24,7 +24,120 @@ const useStyles = makeStyles()(() => ({
     userSelect: 'none',
   },
 }))
+export const HorizontalAxis = observer(function ({
+  model,
+}: {
+  model: DotplotViewModel
+}) {
+  const { viewWidth, borderY } = model
+  const { classes } = useStyles()
+  return (
+    <svg width={viewWidth} height={borderY} className={classes.htext}>
+      <HorizontalAxisRaw model={model} />
+    </svg>
+  )
+})
 
+export const HorizontalAxisRaw = observer(function ({
+  model,
+}: {
+  model: DotplotViewModel
+}) {
+  const { viewWidth, borderX, borderY, hview, htextRotation, hticks } = model
+  const { offsetPx, width, dynamicBlocks, bpPerPx } = hview
+  const dblocks = dynamicBlocks.contentBlocks
+  const hide = getBlockLabelKeysToHide(dblocks, viewWidth, offsetPx)
+  const theme = useTheme()
+  const hviewSnap = {
+    ...getSnapshot(hview),
+    width,
+    staticBlocks: hview.staticBlocks,
+  }
+
+  const ticks = hticks
+    .map(
+      tick =>
+        [
+          tick,
+          bpToPx({
+            refName: tick.refName,
+            coord: tick.base,
+            self: hviewSnap,
+          })?.offsetPx,
+        ] as const,
+    )
+    .filter(f => f[1] !== undefined)
+    .map(f => [f[0], f[1]! - offsetPx] as const)
+
+  return (
+    <>
+      {dblocks
+        .filter(region => !hide.has(region.key))
+        .map(region => {
+          const x = region.offsetPx
+          const y = 0
+          const xoff = Math.floor(x - hview.offsetPx)
+
+          return (
+            <text
+              transform={`rotate(${htextRotation},${xoff},${y})`}
+              key={JSON.stringify(region)}
+              x={xoff}
+              y={y + 1}
+              fill={theme.palette.text.primary}
+              fontSize={11}
+              dominantBaseline="hanging"
+              textAnchor="end"
+            >
+              {region.refName}
+            </text>
+          )
+        })}
+      {ticks.map(([tick, x]) => {
+        return (
+          <line
+            key={`line-${JSON.stringify(tick)}`}
+            x1={x}
+            x2={x}
+            y1={0}
+            y2={tick.type === 'major' ? 6 : 4}
+            strokeWidth={1}
+            stroke={theme.palette.grey[400]}
+          />
+        )
+      })}
+      {ticks
+        .filter(t => t[0].type === 'major')
+        .map(([tick, x]) => {
+          const y = 0
+          return x > 10 ? (
+            <text
+              x={x - 7}
+              y={y}
+              transform={`rotate(${htextRotation},${x},${y})`}
+              key={`text-${JSON.stringify(tick)}`}
+              fill={theme.palette.text.primary}
+              fontSize={11}
+              dominantBaseline="middle"
+              textAnchor="end"
+            >
+              {getTickDisplayStr(tick.base + 1, bpPerPx)}
+            </text>
+          ) : null
+        })}
+      <text
+        y={borderY - 12}
+        x={(viewWidth - borderX) / 2}
+        fill={theme.palette.text.primary}
+        textAnchor="middle"
+        fontSize={11}
+        dominantBaseline="hanging"
+      >
+        {hview.assemblyNames.join(',')}
+      </text>
+    </>
+  )
+})
 export const VerticalAxis = observer(function ({
   model,
 }: {
@@ -54,6 +167,21 @@ export const VerticalAxisRaw = observer(function ({
     width,
     staticBlocks: vview.staticBlocks,
   }
+  const ticks = vticks
+    .map(
+      tick =>
+        [
+          tick,
+          bpToPx({
+            refName: tick.refName,
+            coord: tick.base,
+            self: vviewSnap,
+          })?.offsetPx,
+        ] as const,
+    )
+    .filter(f => f[1] !== undefined)
+    .map(f => [f[0], f[1]! - offsetPx] as const)
+
   return (
     <>
       {dblocks
@@ -77,34 +205,20 @@ export const VerticalAxisRaw = observer(function ({
             </text>
           )
         })}
-      {vticks.map(tick => {
-        const y =
-          (bpToPx({
-            refName: tick.refName,
-            coord: tick.base,
-            self: vviewSnap,
-          })?.offsetPx || 0) - offsetPx
-        return (
-          <line
-            key={`line-${JSON.stringify(tick)}`}
-            y1={viewHeight - y}
-            y2={viewHeight - y}
-            x1={borderX}
-            x2={borderX - (tick.type === 'major' ? 6 : 4)}
-            strokeWidth={1}
-            stroke={theme.palette.divider}
-          />
-        )
-      })}
-      {vticks
-        .filter(tick => tick.type === 'major')
-        .map(tick => {
-          const y =
-            (bpToPx({
-              refName: tick.refName,
-              coord: tick.base,
-              self: vviewSnap,
-            })?.offsetPx || 0) - offsetPx
+      {ticks.map(([tick, y]) => (
+        <line
+          key={`line-${JSON.stringify(tick)}`}
+          y1={viewHeight - y}
+          y2={viewHeight - y}
+          x1={borderX}
+          x2={borderX - (tick.type === 'major' ? 6 : 4)}
+          strokeWidth={1}
+          stroke={theme.palette.grey[400]}
+        />
+      ))}
+      {ticks
+        .filter(t => t[0].type === 'major')
+        .map(([tick, y]) => {
           return y > 10 ? (
             <text
               y={viewHeight - y - 3}
