+++ plugins/dotplot-view/src/DotplotView/components/DotplotView.tsx
@@ -10,8 +10,7 @@ import { DotplotViewModel } from '../model'
 import ImportForm from './ImportForm'
 import Header from './Header'
 import Grid from './Grid'
-import { VerticalAxis } from './VerticalAxis'
-import { HorizontalAxis } from './HorizontalAxis'
+import { VerticalAxis, HorizontalAxis } from './Axes'
 import { TooltipWhereClicked, TooltipWhereMouseovered } from './DotplotTooltip'
 
 const blank = { left: 0, top: 0, width: 0, height: 0 }
