+++ plugins/dotplot-view/src/DotplotView/svgcomponents/SVGDotplotView.tsx
@@ -8,8 +8,7 @@ import { createJBrowseTheme } from '@jbrowse/core/ui'
 // locals
 import { DotplotViewModel, ExportSvgOptions } from '../model'
 import { GridRaw } from '../components/Grid'
-import { HorizontalAxisRaw } from '../components/HorizontalAxis'
-import { VerticalAxisRaw } from '../components/VerticalAxis'
+import { HorizontalAxisRaw, VerticalAxisRaw } from '../components/Axes'
 import SVGBackground from './SVGBackground'
 
 // render LGV to SVG
